lsm/io: remove dead cloud_persistence code
The cloud_persistence code is now superceded with
cloud_cache_persistence. Updates the remaining callers (just tests) to
use the cloud_cache_persistence instead of cloud_persistence, and moves
the helpers and shared code into cloud_cache_persistence.

(cherry picked from commit f4781ff7462e17069e5879ac128ef6edb2cfdf41)
