+++ src/v/cloud_topics/read_replica/tests/BUILD
@@ -16,7 +16,7 @@ redpanda_test_cc_library(
         "//src/v/cloud_topics/level_one/metastore/lsm:write_batch_row",
         "//src/v/container:chunked_hash_map",
         "//src/v/lsm",
-        "//src/v/lsm/io:cloud_persistence",
+        "//src/v/lsm/io:cloud_cache_persistence",
         "//src/v/model",
         "//src/v/ssx:time",
         "@seastar",
@@ -32,13 +32,16 @@ redpanda_cc_gtest(
     cpu = 1,
     deps = [
         ":db_utils",
+        "//src/v/cloud_io:cache",
         "//src/v/cloud_io:remote",
         "//src/v/cloud_io/tests:s3_imposter",
         "//src/v/cloud_io/tests:scoped_remote",
         "//src/v/cloud_topics/level_one/metastore:domain_uuid",
         "//src/v/cloud_topics/read_replica:snapshot_manager",
         "//src/v/cloud_topics/read_replica:snapshot_metastore",
+        "//src/v/config",
         "//src/v/lsm",
+        "//src/v/storage:disk",
         "//src/v/test_utils:gtest",
         "//src/v/test_utils:scoped_config",
         "//src/v/test_utils:tmp_dir",
